Reorganization, comments
