+++ crates/tracing/src/stack.rs
@@ -1,4 +1,10 @@
-// use std::fmt;
+//! This module is centered around a thread-local Stack of Span references,
+//! with some public functions for pushing and accessing values from the stack.
+//! This is primarily used by the autotrace proc macro in holochain_tracing_macros,
+//! to automatically push new child spans onto the stack when entering a new function frame.
+//! Using functions like `with_top` allow the user to access the stack directly, for situations
+//! like needing to take a span context and send it into another thread, or out of the process entirely.
+
 use std::cell::RefCell;
 use std::rc::Rc;
